move connecting widget to utils
