+++ packages/scaffold/index.ts
@@ -21,7 +21,6 @@ export * from './src/partials/w3m-connecting-wc-mobile/index.js'
 export * from './src/partials/w3m-connecting-wc-qrcode/index.js'
 export * from './src/partials/w3m-connecting-wc-unsupported/index.js'
 export * from './src/partials/w3m-connecting-wc-web/index.js'
-export * from './src/partials/w3m-connecting-widget/index.js'
 export * from './src/partials/w3m-header/index.js'
 export * from './src/partials/w3m-help-widget/index.js'
 export * from './src/partials/w3m-snackbar/index.js'
