+++ packages/scaffold/src/partials/w3m-connecting-wc-mobile/index.ts
@@ -1,6 +1,6 @@
 import { ConnectionController, CoreHelperUtil } from '@web3modal/core'
 import { customElement } from 'lit/decorators.js'
-import { W3mConnectingWidget } from '../w3m-connecting-wc-super/index.js'
+import { W3mConnectingWidget } from '../../utils/w3m-connecting-widget/index.js'
 
 @customElement('w3m-connecting-wc-mobile')
 export class W3mConnectingWcMobile extends W3mConnectingWidget {
