--- packages/scaffold/src/partials/w3m-connecting-widget/styles.ts
@@ -1,61 +1,0 @@
-import { css } from 'lit'
-
-export default css`
-  @keyframes shake {
-    0% {
-      transform: translateX(0);
-    }
-    25% {
-      transform: translateX(3px);
-    }
-    50% {
-      transform: translateX(-3px);
-    }
-    75% {
-      transform: translateX(3px);
-    }
-    100% {
-      transform: translateX(0);
-    }
-  }
-
-  wui-flex:first-child:not(:only-child) {
-    position: relative;
-  }
-
-  wui-loading-thumbnail {
-    position: absolute;
-  }
-
-  wui-icon-box {
-    position: absolute;
-    right: calc(var(--wui-spacing-3xs) * -1);
-    bottom: calc(var(--wui-spacing-3xs) * -1);
-    opacity: 0;
-    transform: scale(0.5);
-    transition: all var(--wui-ease-out-power-2) var(--wui-duration-lg);
-  }
-
-  wui-text[align='center'] {
-    width: 100%;
-    padding: 0px var(--wui-spacing-l);
-  }
-
-  [data-error='true'] wui-icon-box {
-    opacity: 1;
-    transform: scale(1);
-  }
-
-  [data-error='true'] > wui-flex:first-child {
-    animation: shake 250ms cubic-bezier(0.36, 0.07, 0.19, 0.97) both;
-  }
-
-  [data-retry='false'] wui-button {
-    display: none;
-  }
-
-  [data-retry='true'] wui-button {
-    display: block;
-    opacity: 1;
-  }
-`
