+++ packages/scaffold/src/views/w3m-connecting-external-view/index.ts
@@ -1,6 +1,6 @@
 import { ConnectionController, ModalController } from '@web3modal/core'
 import { customElement } from 'lit/decorators.js'
-import { W3mConnectingWidget } from '../../partials/w3m-connecting-wc-super/index.js'
+import { W3mConnectingWidget } from '../../utils/w3m-connecting-widget/index.js'
 
 @customElement('w3m-connecting-external-view')
 export class W3mConnectingExternalView extends W3mConnectingWidget {
